Leaf 4042 - remove double unlink in mailer

+++ LEAF_Request_Portal/mailer/mailer.php
@@ -56,13 +56,14 @@ foreach ($queue as $item)
     if (strlen($item) == 40)
     {
         // attempt to resend email if its 5 minutes old
-        if (time() - filemtime($folder . $item) >= 300)
+        if (file_exists($folder . $item) && time() - filemtime($folder . $item) >= 300)
         {
             $email = unserialize(file_get_contents($folder . $item));
             if (strlen(trim($email['recipient'])) == 0)
             {
                 // delete invalid cache
                 unlink($folder . $item);
+
                 trigger_error('Mail no recipient: ' . $email['subject']);
             }
             else
